Refacto to remove some code duplication across staking flows
