+++ src/app/stake/StakingPositionSelector.tsx
@@ -30,109 +30,6 @@ type StakingPositionSelectorProps = {
   onSelect: (stakingPosition: StakingPosition, index: number) => void;
 };
 
-const StakingPositionView = ({
-  stakingPosition,
-  validators,
-}: {
-  stakingPosition: StakingPosition;
-  validators: Validator[];
-}) => {
-  // FIXME Hack for Cosmos, all validatorAddresses should be handled not just the 1st one
-  const validator = useMemo(
-    () =>
-      stakingPosition &&
-      validators.find(
-        (v) => v.address === stakingPosition?.validatorAddresses[0]
-      ),
-    [stakingPosition, validators]
-  );
-
-  return (
-    validator && (
-      <div className="flex items-center justify-between w-full">
-        {validator.name && (
-          <div className="relative">
-            <Tooltip text={validator?.address}>
-              <TooltipTrigger>
-                <Avatar className="w-[32px] h-[32px]">
-                  <AvatarFallback>
-                    {validator.name[0].toUpperCase() ||
-                      validator?.address[0].toUpperCase()}
-                  </AvatarFallback>
-                </Avatar>
-              </TooltipTrigger>
-            </Tooltip>
-            {validator.chainLogo && (
-              <Tooltip text={validator.chainId}>
-                <TooltipTrigger>
-                  <div className="absolute w-4 h-4 text-xs font-bold text-primary bg-primary-foreground border-2 rounded-full -top-[6px] -end-1">
-                    <Avatar className="h-3 w-3">
-                      <AvatarImage
-                        src={validator.chainLogo}
-                        alt={validator.chainId}
-                      />
-                      <AvatarFallback>{validator.chainId}</AvatarFallback>
-                    </Avatar>
-                  </div>
-                </TooltipTrigger>
-              </Tooltip>
-            )}
-          </div>
-        )}
-        <div className="flex-1 text-right">{validator.name}</div>
-        <div className="font-bold flex-1 text-right">
-          Stake: {stakingPosition.amount}
-          {/* FIXME Need to display the unit too */}
-        </div>
-      </div>
-    )
-  );
-};
-
-const StakingPositionSelectorList = ({
-  setOpen,
-  setSelectedChoice,
-  onSelect,
-  stakingPositions,
-  validators,
-}: {
-  setOpen: (open: boolean) => void;
-  setSelectedChoice: (choice: StakingPosition | undefined) => void;
-  onSelect: (stakingPosition: StakingPosition, index: number) => void;
-  stakingPositions: StakingPosition[];
-  validators: Validator[];
-}) => {
-  return (
-    <Command>
-      <CommandInput placeholder="Filter positions..." />
-      <CommandList>
-        <CommandEmpty>No results found.</CommandEmpty>
-        <ScrollArea className="h-[240px] overflow-auto">
-          <CommandGroup>
-            {stakingPositions.map((stakingPosition, i) => (
-              <CommandItem
-                key={`${stakingPosition.validatorAddresses}_${i}`}
-                value={`${stakingPosition.validatorName}_${i.toString()}`}
-                onSelect={(value) => {
-                  const [name, index] = value.split("_");
-                  setSelectedChoice(stakingPositions[Number(index)]);
-                  setOpen(false);
-                  onSelect(stakingPosition, i);
-                }}
-              >
-                <StakingPositionView
-                  stakingPosition={stakingPosition}
-                  validators={validators}
-                />
-              </CommandItem>
-            ))}
-          </CommandGroup>
-        </ScrollArea>
-      </CommandList>
-    </Command>
-  );
-};
-
 export function StakingPositionSelector({
   stakingPositions,
   validators,
@@ -191,7 +88,7 @@ export function StakingPositionSelector({
               validators={validators}
             />
           ) : (
-            <>Select a validator</>
+            <>Select a position</>
           )}
         </Button>
       </DrawerTrigger>
@@ -209,3 +106,107 @@ export function StakingPositionSelector({
     </Drawer>
   );
 }
+
+const StakingPositionSelectorList = ({
+  setOpen,
+  setSelectedChoice,
+  onSelect,
+  stakingPositions,
+  validators,
+}: {
+  setOpen: (open: boolean) => void;
+  setSelectedChoice: (choice: StakingPosition | undefined) => void;
+  onSelect: (stakingPosition: StakingPosition, index: number) => void;
+  stakingPositions: StakingPosition[];
+  validators: Validator[];
+}) => {
+  return (
+    <Command>
+      <CommandInput placeholder="Filter positions..." />
+      <CommandList>
+        <CommandEmpty>No results found.</CommandEmpty>
+        <ScrollArea className="h-[240px] overflow-auto">
+          <CommandGroup>
+            {stakingPositions.map((stakingPosition, i) => (
+              <CommandItem
+                key={`${stakingPosition.validatorAddresses[0]}_${i}`}
+                value={`${stakingPosition.validatorName}_${i.toString()}`}
+                onSelect={(value) => {
+                  const [name, index] = value.split("_");
+                  setSelectedChoice(stakingPositions[Number(index)]);
+                  setOpen(false);
+                  onSelect(stakingPosition, i);
+                }}
+              >
+                <StakingPositionView
+                  stakingPosition={stakingPosition}
+                  validators={validators}
+                />
+              </CommandItem>
+            ))}
+          </CommandGroup>
+        </ScrollArea>
+      </CommandList>
+    </Command>
+  );
+};
+
+const StakingPositionView = ({
+  stakingPosition,
+  validators,
+}: {
+  stakingPosition: StakingPosition;
+  validators: Validator[];
+}) => {
+  const validator = useMemo(() => {
+    // FIXME Hack for Cosmos, all validatorAddresses should be handled not just the 1st one
+    return (
+      stakingPosition &&
+      validators.find(
+        (validator) =>
+          validator.address === stakingPosition.validatorAddresses[0]
+      )
+    );
+  }, [stakingPosition, validators]);
+
+  return (
+    validator && (
+      <div className="flex items-center justify-between w-full">
+        {validator.name && (
+          <div className="relative">
+            <Tooltip text={validator?.address}>
+              <TooltipTrigger>
+                <Avatar className="w-[32px] h-[32px]">
+                  <AvatarFallback>
+                    {validator.name[0].toUpperCase() ||
+                      validator?.address[0].toUpperCase()}
+                  </AvatarFallback>
+                </Avatar>
+              </TooltipTrigger>
+            </Tooltip>
+            {validator.chainLogo && (
+              <Tooltip text={validator.chainId}>
+                <TooltipTrigger>
+                  <div className="absolute w-4 h-4 text-xs font-bold text-primary bg-primary-foreground border-2 rounded-full -top-[6px] -end-1">
+                    <Avatar className="h-3 w-3">
+                      <AvatarImage
+                        src={validator.chainLogo}
+                        alt={validator.chainId}
+                      />
+                      <AvatarFallback>{validator.chainId}</AvatarFallback>
+                    </Avatar>
+                  </div>
+                </TooltipTrigger>
+              </Tooltip>
+            )}
+          </div>
+        )}
+        <div className="flex-1 text-right">{validator.name}</div>
+        <div className="font-bold flex-1 text-right">
+          Stake: {stakingPosition.amount}
+          {/* FIXME Need to display the unit too */}
+        </div>
+      </div>
+    )
+  );
+};
